update tools/x2coco.py shell
line 193 was " labels_str = f.read().split()", this will lead to a wrong result when use it for voc dataset transfer to coco, because the image labels in the file are segment by '\n" when wrote to the file. if use a default seg(" "）， when the label is like "haski dog", will result to error!