Make LCPAN also pass the activator to DepthwiseSeparable
Without this, nn.Hardswish() will be selected in ConvBNLayer, which causes the opset version to be bumped to 14, in turn breaking a ESP-DL onnx quantization, which only  supports 13. 

I would also suspect that this is more correct also in other cases as one would not be likely to not want to use a specific activation function everywhere, would one supply it. 

Also, it aligns with how the act-parameter is handled in other cases.